Adding commands to create/save threads and convos
This will allow us to recreate conversations that got spread out within a channel, and either save it or create a new thread from it. Once we have a thread, the new commands will also allow us to save that thread easily.

I've also added commands to allow us to list the saved conversations in a channel, and to load one for download if we want.

Saved conversations are in Markdown format, so we should be able to render them to HTML.

TODO: Convert to slash commands to give a more interactive experience.
TODO: Save conversations in YAML to give more rendering flexibility.